Make NoopScope metric scope really a no-op operation
When workflow client code don't provide a metric scope, a default "NoopScope"
is instantiated by cadence client framework. With the current implementation of
NoopScope, a single threaded executor is indirectly created by RootScopeBuilder
for reporting empty metrics at regular interval. This executor is never shutdown.

The side effect is that a client process that just launches a workflow and returns
from main method will stay alive due to the non daemon thread of this executor.

NoopScope rewritten to not launch any reporting threads.