Implement activity helpers
 As a developer I would like to be able to define or even implements
 commons methods (default methods)  that could be extended by an
 activity method interface. This methods, should not be consider as an
 activity method.

This feature will gives an upper level of abstraction allowing commons
behavior between activity methods.

In order to support this feature, an annotation named 'ActivityHelper'
was developed. You must add this annotation to all those interfaces that
gives some kind of support or gives an upper level of abstraction to an
activity method interface.