Fix prometheus reporting issue
MetricsType.ACTIVITY_SCHEDULED_TO_START_LATENCY is reported twice with different set of tags.
It causes prometheus to fail with an IllegalArgumentException: "Prometheus requires that all meters with the same name have the same set of tag keys."
Made all metrics to be reported with the same set of tags.

Improved reporter to track gauges as well.