Fix Promise#allOf(Promise<?>...) (#881)
AllOfFuture completes with the list of values from each promise passed into its constructor. The varargs variant of Promise#allOf however has a return type of Void, so the intention seems to be to discard the results of the promises. Because a list of results is disjoint from Void any invocation of this method will throw an exception when the promises complete.

To resolve this we explicitly discard the result of the promises by mapping the result to null, the only possible value of Void.