Fix flakiness in LoggerTest (#885)
If any other threads logs a message while we iterate over the events it can throw a ConcurrentModificationException.

Create our own implementation of ListAppender that uses the synchronized keyword to protect reading/writing to the backing list with a mutex. Create a copy of the list while holding to the lock to support iterating over it.