Remove Reflection from CheckedExceptionWrapper
CheckedExceptionWrapper#unwrap currently uses reflection to rewrite the chain of Exceptions to remove all instances of CheckedExceptionWrapper. This approach is nonstandard to how Exception wrapping usually works in Java, and isn't friendly with the encapsulation in Java 17. Rather than trying to work around it, remove the behavior.

RpcRetryer was particularly impacted because both it and CompletableFuture wrap exceptions. Refactor it to avoid wrapping them using CheckedExceptionWrapper so that the final exception remains the same.